Change references to qrc file in *.ui files
The application.qrc file with icons has been moved to the src-common
directory, but the *.ui files are still referenced in the old
location. This did not cause any errors or warnings, but did cause
some inconvenience when opening and editing the *.ui files.

The *.rc files, which are no longer in use, have also been removed.